refactor measurable switch into non-overlapping module
